fix: cache overlay actors, fix stale index, and extract shared preset logic
- Cache overlay St actors with monitor fingerprinting instead of
  recreating on every workspace switch
- Update workspace index before the enabled check to prevent stale
  overlay trigger on re-enable
- Add destroyed guard to prevent callbacks after cleanup
- Extract duplicated preset definitions and logic from BehaviorPage
  and AppearancePage into shared overlayPresets module
